Split test file into multiple test files
